requirements.txt
fastapi==0.109.0
pydantic==2.6.0
pyserial==3.5
python-dotenv==1.0.1
